Move registry ingress resources out of nginx ingress controller service

+++ plugins/extra-kind-resources/index.sh
@@ -15,10 +15,25 @@ if [ "$COMMAND" == "install" ]; then
 
   echo "$CONFIG" | yq -p=json > .values.yaml;
 
-  # delete this resource, so we can add it anew with modifications
-  kubectl delete -n kube-system configmap coredns || true;
+  skaffold run 2> /dev/null;
+  rm .values.yaml
+fi
 
+if [ "$COMMAND" == "install" ]; then
+  CLUSTER_NAME=""
+  CONFIG=""
+
+  for arg in "$@"; do
+    if [[ $arg == --clusterName=* ]]; then CLUSTER_NAME="${arg#--clusterName=}"; fi
+    if [[ $arg == --config=* ]]; then CONFIG="${arg#--config=}"; fi
+  done
+
+  # delete these resources, so we can add it anew with modifications
+  kubectl delete -n kube-system configmap coredns || true;
   skaffold delete || true;
+
+  echo "$CONFIG" | yq -p=json > .values.yaml
+  yq -i ".clusterName = \"$CLUSTER_NAME\"" .values.yaml
   skaffold run 2> /dev/null;
   rm .values.yaml
 fi
